requirements.txt
streamlit
streamlit-chat
openai==0.28
pinecone-client
tiktoken
pinecone[grpc]
googleapis-common-protos
protoc_gen_openapiv2
tiktoken 
poppler-utils  
unstructured 
unstructured[local-inference] 
# detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.6#egg=detectron2 
sentence-transformers
google-api-python-client
grpcio
pypdf
langchain_pinecone
google-api
langchain
langchain-community 
langchain-core 
langchain-openai 
langchain-pinecone
langchain-text-splitters
pypdf                   
grpcio
 